+++ src/main/java/de/igslandstuhl/database/Application.java
@@ -10,7 +10,7 @@ import de.igslandstuhl.database.api.Topic;
 import de.igslandstuhl.database.holidays.Holiday;
 import de.igslandstuhl.database.server.Server;
 import de.igslandstuhl.database.server.commands.Command;
-import de.igslandstuhl.database.server.webserver.PostRequestHandler;
+import de.igslandstuhl.database.server.webserver.handlers.PostRequestHandler;
 import de.igslandstuhl.database.utils.CommandLineUtils;
 
 /**
