+++ src/main/java/de/igslandstuhl/database/server/webserver/handlers/HttpHandler.java
@@ -1,9 +1,12 @@
-package de.igslandstuhl.database.server.webserver.requests;
+package de.igslandstuhl.database.server.webserver.handlers;
 
 import de.igslandstuhl.database.Registry;
 import de.igslandstuhl.database.server.Server;
 import de.igslandstuhl.database.server.webserver.AccessLevel;
 import de.igslandstuhl.database.server.webserver.Status;
+import de.igslandstuhl.database.server.webserver.requests.APIPostRequest;
+import de.igslandstuhl.database.server.webserver.requests.GetRequest;
+import de.igslandstuhl.database.server.webserver.requests.HttpRequest;
 import de.igslandstuhl.database.server.webserver.responses.HttpResponse;
 import de.igslandstuhl.database.server.webserver.sessions.SessionManager;
 import de.igslandstuhl.database.utils.ThrowingFunction;
