+++ src/test/java/de/igslandstuhl/database/server/webserver/AccessManagerTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Test;
 
 import de.igslandstuhl.database.api.User;
 import de.igslandstuhl.database.server.resources.ResourceLocation;
+import de.igslandstuhl.database.server.webserver.handlers.WebResourceHandler;
 
 public class AccessManagerTest {
     private User teacher;
